Move FileComponents into its own class
